Give k8s-infra-prow-build cluster time to scale-up
We have k8s-infra-prow-build etup to autoscale if it doen't have
enough capacity. Anecdotally I saw it take 5sec to decide that
it needed to scale, but another ~2min for a node to come online.

Let's assume it could take longer, so wait up to 5min before calling
an unscheduled pod safe to delete.