Raise ci-kubernetes-build-fast timeout to 40min
Looking at recent history it is not uncommon for it to timeout
N times in a row when it's actually building release artifacts.

Since this has been happening before we put resource limits in
place, I'm inclined to say the timeout is too tight. Let's raise
it to 40min before we consider throwing more cpu at it