Add kube-up/down and dump logs dependencies to KRTE
The recently-released GCE deployer for kubetest2 will be run
in KRTE containers in CI. KRTE is currently missing a couple of
necessary dependencies:

The deployer calls out to kube-up.sh, which needs gcloud alpha
and beta components.

The deployer also uses a log dumping script which relies on
gcloud compute ssh and scp. This adds a necessary SSH client.