Update plank dashboard to filter and group by more things
Allow someone to filter and group by:
- cluster
- org
- repo
- state
- type

Add a third graph panel, and change each graph panel to group by
user-selectable labels

I intially had a graph panel for each of these, but:
- it's a lot of clutter
- it appeared to be timing out prometheus at the 90d range

It turns out each template re-queries prometheus (serially) anytime the
dashboard's time-range changes, and only after templates are refreshed do
queries fired off for the dashboard's panels in parallel. This is necessary
because the `label_values` prometheus query is time-range dependent.

To mitigate this I changed the "type" and "state" templates to use a hardcoded
list of values, since I think it unlikely we're going to add new prowjob types
or states.