trigger: avoid double-triggering on `/ok-to-test
When someone comments `/lgtm`, triggers reacts to that comment by 1)
adding the label 2) triggering the tests. Adding the label causes
another event to happen, which the logic attempts to skip by checking
that the event was caused by the bot, but it incorrectly used the author
(of the PR) for this comparison, not the `sender` field, which is the
account that caused the event to happen.

This PR fixes the comparison to use the event `sender` field.