Prepare config-forker for unique names in prow
Create unique names on forked jobs and modify the reporting context in
an expected way. If the original context did not include `-master`, keep
it in the original state, to not invent new context values on release
branches. Otherwise replace -master with the version to make the github
context distinct from the master jobs.

Signed-off-by: Roman Mohr <rmohr@redhat.com>