Crier github reporter: Don't block workers waiting for lock
In 00562873e0c597be9d186ebc3887c3fa0690400c we introduced PR-Level
locking for presubmits to the crier github reporter in order to allow
running it with multiple workers without having the workers race with
each other when creating the GitHub comment.

The implementation naively used a map of sync.Mutex, blocking workers
that are waiting for their lock. This can result in all workers being
blocked, which is worse than the initial problem of only having one
worker.

This commit updates that to instead use a semaphore.Weightened and
TryAcquire, returning a RequeueAfter if the lock is occupied in order to
minimize lock contention.