Revert tagMatch flag for image builder
This flag was introduced in https://github.com/kubernetes/test-infra/pull/18506 and
later attempted to be fixed in https://github.com/kubernetes/test-infra/pull/18540. However,
this command won't work with Go's cmd libraries as this Git flag takes a `glob`, which are
not expanded by Go.

While this would be a nice feature to have, we should continue searching for a way to implement it.
Though in order to get image builds working for the Descheduler again, we need to remove
it for now.