Bumper: Allow user to specify working directory for git push (#19159)
* Add GetAllClones and GetRoot functions

These functions have been added as part of the bugzilla backporter tool. 
The GetAllClones function would return all the clones of a particular bug(including its parent, grandparent etc.), unlike the GetClones function which would return only the clones which were cloned from that particular bug.
The GetRoot function would return the original bug which produced the clone even if that bug is a clone of a clone.

* Add login and name options for git commits

* Move git author options to bumper.go

* Use constant instead of hardcoding "git"