include unittests for the gcsweb tool, introduce the gcs mock server
This commit includes unittests for the gcsweb tool. In order to make
sure that the calls using the cloud storage client are correct, and
since there isn't any other way to test it with any fake client, this
commit introduces a small gcs mock server with a trivial functionality
that the tool already uses.

Signed-off-by: Nikolaos Moraitis <nmoraiti@redhat.com>