Adds windows servercore cache image periodic job
The base image for the Windows E2E test images is nanoserver. In most cases,
that is sufficient. But in some cases, we are missing some DLLs.

We can fetch those DLLs from Windows servercore images, but they are very large
(2GB compressed), while the DLLs are only a few megabytes in size. We can build
a monthly DLL cache image and use the cache instead.