Allow staging via krel for kubetest
This adds a new `--use-krel` flag to kubetest to indicate that we want
to stage the sources directly via the API provided by k/release rather
than the deprecated push-build.sh script (lives in k/release, too).

The idea is to add another job later on for testing this flag on a non
critical Kubernetes build+push job (should live in sig-release).

Signed-off-by: Sascha Grunert <sgrunert@suse.com>
Signed-off-by: Sascha Grunert <mail@saschagrunert.de>