Add deletion annotation handling for job forking
The `fork-per-release-deletions` annotation can be used for custom deletions in
your `labels` (periodic jobs only). This is a comma-separated list of keys of
labels you would like to remove on forking, for instance:

```
fork-per-release-deletions: "label-key-to-delete1, label-key-to-delete2"
```

This is useful for getting rid of a specific preset in the forked job. For
instance, one can have a master branch job that has a label corresponding to a
master-specific preset which is undesired for forked release jobs.