sig-storage images: periodically rebuild csi-driver-host-path
We are using the "canary" images from the staging repo for some tests.
If we don't push changes into a repo for a while, there's a risk that
the existing canary image from the post-submit job becomes so old that
it gets removed.

Periodically rebuilding the image avoids that. It also ensures that we
detect issues like not being compatible with the cloud build service
before we really need it.