releng(gcr-backup): Use k8s.gcr.io/releng/releng-ci:v0.1.1
The canary job mimics 'ci-k8sio-backup', which is using the following
image:

gcr.io/k8s-testimages/kubekins-e2e:v20201031-122dc79-1.17

A few things wrong with this:
- kubekins is a bit more bloated than we need for this use case
- the kubekins 1.17 image variant is being used, which is incorrect

This switches to using a versioned and promoted lightweight image, which
is maintained by Release Engineering.

An additional advantage here is that we (Release Managers) are in full
control of the image and can turn around any required changes faster
than the kubekins-e2e image.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>