releng: Teach jobs using latest-fast marker to extract from K8s Infra
As we continue to migrate release-blocking jobs to a dedicated K8s Infra
cluster, jobs that use the latest-fast marker need to extract builds
from gs://k8s-release-dev, which is the K8s Infra equivalent of
gs://kubernetes-release-dev.

A new flag ('--extract-ci-bucket=k8s-release-dev') was added to support
this transitional use case, so we employ it here.

Exceptions:
Of note, the ci-kubernetes-e2e-gce-master-new-gci-kubectl-skew-serial is
not included in this PR.

This job does two extractions:
- --extract=ci/k8s-stable1
- --extract=ci/latest-fast

As the generic version markers (like 'k8s-stable1') have not been
migrated to K8s Infra, we cannot take advantage of this flag.

We'll plan to fix this job in a follow-up.

Signed-off-by: Stephen Augustus <saugustus@vmware.com>