kubernetes-csi: set snapshotter version
Some jobs need the "master" version of the snapshotter CRD and
controller. For others it makes sense to select the version that gets
tested in the Prow job instead of having to bump the prow.sh script in
all affected repos.

The caveat is that all future versions of external-snapshotter must
install through exactly the YAML files that are currently used by
prow.sh:
https://github.com/kubernetes-csi/csi-release-tools/blob/5d874cce4e649dfd254d01b9b44179ffa72aee75/prow.sh#L702-L722