Run master-informing capg against release-1.19
This is a temporary fix to make the capg master-informing job run
against the release-1.19 conformance tests while it is using the latest
stable kubernetes version from 1.19.

This should be followed by investigating whether CAPI jobs should be on
any master dashboards for testgrid when they are running against older
branches.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>