Github v4: Add apps auth support
This change adds apps auth support to the Github V4 client. This is
achieved by adding a new method that takes an additional org argument,
then using the context to pass this through to the underlying
roundtripper. For consistency, the client was changed to use the Context
for this everywhere instead of a header on the V3 api.

We use the context to pass the org on because the code that constructs
the V4 requests is part of a dependency, giving us no way to use a
Header. (Ab-)using the context for this allows us to get away without
requiring changes to our dependencies.