test images: Adds dockerhub images used in tests
Adds the prow config jobs for images that are currently hosted on dockerhub
but are being used in E2E tests. The image Dockerfiles are mostly noop since
the purpose of these jobs is to mirror those images into our own registries,
so we can then use them in tests instead of the dockerhub ones.