Fix self-detection for github reporting when using apps auth
This change fixes the selfdetection in the github reporting logic when
using apps auth, where the comment author will be suffixed with [bot].

To do so, a checker function can be requested from the github client
that accounts for this. This approach was taken over directly exposing a
method on the github client for this because such a method would need to
also return an error. This in turn would mean that we would have to
change a lot of logic throughout the codebase where we currently just
pass a string around and do not always have the possibility to return an
error.