resource request for network policy job
the periodic job that test network policies is not stable, random
test fail but we couldn't observe any pattern.

Some of the test fail because the endpoints are not ready, to discard
first that we have a possible problem with resources constraints, we
allocate memory to the nodes, matching the current presubmit job.