Crier, Deck, Prow-CM, Tide: Restart on kubeconfig changes
This change makes Crier, Deck, the Prow Controller Manager and Tide exit
on kubeconfig changes to make the Kubelet restart them in order to pick
up those changes.

The underlying funtionality was added and enabled in Sinker a couple of
weeks back in 3b2ff68d029fce94fd9c1fbd3651b6427db34219 and seems to be
working fine.