prow: allow configuring the filenames used for OWNERS(_ALIASES)
When a repository is a fork of an upstream repository with OWNERS, or
vendors a repository that also uses OWNERS, it is generally not intended
for users who are OWNERS upstream to have rights in the downstream.
Furthermore, changing the OWNERS files leads to merge conflicts as
changes from upstream are pulled in, which leads to headaches. Allowing
users to configure the files used by Prow to denote owners and aliases
side-steps this issue.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>