Config: Insert git sha when using configupdater
This change makes the configupdater always add a key named VERSION to
the configmap which contains the git sha that triggered said update.
The configloading in turn will pick it up if present. This allows
components to include the config version in their logs, which can be
useful for debugging.