Mkpj doesn't fetch github when running at local mode
One of the use case of mkpj is to couple with phaino, when --local is passed in as param. Since phaino utilized pre-checked out source code, instead of source code clone or checkout, there is no need to fetch github for a presubmit or postsubmit job.

This change also makes it possible for source control system other than github to use mkpj + phaino for running prow jobs locally