Fix cri-o download urls
cri-o use 'git rev-parse --short HEAD' as the filename, and the length
of the sha can vary. (in theory from 4 to 40 chars, but less in
practice). The output can vary between machines due to config, so it is
not stable in that sense.

This tries with length 7, 8 and 9, and works as a temporary solution to
the problem. Will connect to the cri-o people to see if they can come up
with a stable solution.