Update ci-k8sio-audit periodic job github token
This also adds dynamically pulling the email and name via the token
rather than needing to hard code them in the job.

Note that it's the default email that is used for the token provided.

Also switched to parsing curl output via jq rather than sed. GNU sed and
BSD sed act a bit differently and I was getting inconsistent results on
OSX vs Linux.