prow/plugins: allow label-based automatic posting of comments
This commit allows automatically posting a comment based on the presence
of certain labels (labels matching a particular regexp).

Some examples where this feature can be useful:

1. If an issue contains the `kind/feature` label, the bot can add a
comment pointing to the KEPs process.

2. If an issue contains the `kind/support` label, the bot can add a
comment pointing to the support resources. The bot can also automatically
close the issue with a `/close` comment.

This feature is being added to the `require-matching-label` plugin.
While this feature isn't inherently related to requiring a particular
label, porting this feature to a new/different plugin requires
duplication of code.