Prow dashboard: use job name instead of ip address for filtering boskos instances
There are prow instance(s) running boskos, the metrics it emits are from the pod running boskos, the instance name (ip address) changes when the boskos pod is recreated, which makes it impossible to use it as diffentiator. Use job name as differentiator is more reliable and less easy breaking when external boskos ip changes.

Note: this is a breaking change, boskos dashboard will only display values from k8s-infra-prow-builds-boskos after this PR