Fix HNC postsubmits
Our first postsubmits failed, apparently because Prow changes the
workdir for postsubmits but not periodics. This change will work with
https://github.com/kubernetes-sigs/multi-tenancy/pull/1398 (which is not
yet submitted, but whose image has already been pushed) to ensure that
the e2e tests work as both periodic and postsubmits.

Tested: cannot test prow; as soon as the PR mentioned above is merged,
we'll find out if the postsubmits work or not.