secrets: censor more efficiently
The `bytes.ReplaceAll()` call will copy the input, so the previous
implementation of the secret agent's censoring would take up O(MN)
memory where M is the size of the input and N is the number of secrets
that are loaded. There's no need for that. The `bytereplacer` utility
will operate in-place (unless it needs to reallocate for a longer
replacement) and is faster, as well. I've kept the old `Censor()`
function around and it will continue not to mutate input. The content of
the censored text is changing, though, from the literal `"CENSORED"`
string to repeated `"*"` characters as long as the input.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>