prow: Add ignore-space-change flag to git am
This commit adds the ignore-space-change flag to git am calls in the git
client. This is being done as certain patches are not able to applied.
This is unexpected behavior as the commits the patches have been created
from are able to be cherry-picked. This behavior may be caused by
carriage returns being present in the contents of the patch.

Signed-off-by: Sebastian Soto <ssoto@redhat.com>