Sinker: Instantly clean duplicate prowjobs
This change adds a feature gated addition to sinker to make it delete
duplicate prowjobs. Prowjobs that are not periodics, have finished and
have the same name, refs, extrarefs and result are considered
duplicates.

The reason for wanting this is to allow increasing the prowjob retention
without overloading the api. The prowjob retention is important,
because we always have to repeat tests for prs after max_prowjob_age.