sidecar: create symlinks dynamically
Bazel will de-reference symlinks before starting the build sandbox, so
the only way we can get them to exist at the time that we run the test
is by creating them at runtime. Place files named `link` into the
directory where the content is the location the link should point to.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>