Tide: Prioritize batches with pre-existing tests when picking
Currently, tide will always pick the biggest possible batch. This is
suboptimal in high-traffic repos, because there is a good chance that
new PRs become merge-eligible in the timeframe it takes for tests to
fail. This means that we will have to completely start over, even if
only a single test failed in the previous attempt.

This change changes that logic to prioritize batches by number of past
successful job runs or if none exists by number of currently pending
job runs. The rationale is that:
* If we have past successful runs and assume all jobs have a non-zero
  flakerate, the likelihood of succeeding is increased
* Tide will omit the creation of jobs that already exist and are not
  failed, so this reduces the overall number of jobs we run
* If there is already an existing pending job, it is guaranteed to
  finish quicker than a new job for a new batch

The main argument against this change is that it means that we continue
testing smaller batches when a bigger one is available. In the grand
scheme of things this doesn't really make a difference though, because
if a new batch n+1 is bigger than the existing n batch, the batch
following that (n+2) will be smaller by the same number of PRs batch n+1
was bigger than n.

A nice side-effect is that this makes reasoning easier when a subset of
jobs consistently fails (i.E. due to PRs in the batch being incompatible
or the job being broken) because one doesn't have to find the subset of
jobs that fails consistently in different batches anymore.