peribolos: return more useful errors
Previously the errors propagated to the final summary printed the actual
errors by GH API, but not what operations were actually attempted. So in
the usual case where one or two operations fail you needed to scroll and
search in the log.

Include the attempted operation in the error so that this information
gets propagated to the summary.