Separate scheduler-plugins CI Jobs For Each Release Branch
Each release branch in the scheduler-plugins repo is using a different
version of Go. This change creates a separate set of CI jobs for each
release branch, so that the Go version for each branch can be updated
and tested separately.

Also, bumps the Go version for the master branch from 1.15.8 to
1.15.10.

This is modeled after the CI jobs for the descheduler:
https://github.com/kubernetes/test-infra/tree/master/config/jobs/kubernetes-sigs/descheduler