Crier gerrit reporter: skip reporting if job from newer patchset already exists
In a gerrit PR, when a new change is pushed to it the previous batch of presubmit jobs were aborted, if all jobs haven't finished by the time of abortion then gerrit won't report because aborted state wouldn't trigger a report. But if there was a single job failed before the abortion then crier wouldn't report the failed prowjob because by the time of failure the other jobs were still running, this is also expected. However, lately in prow serving gerrit repos, these earlier failed jobs were reported long time later, in the range of 3-12 hours according to my observation. It perhaps is related to prow bump that restarted crier which makes it reconcile all prowjobs and triggered the reporting of prowjob failed long time ago.

This PR is fixing this problem by stopping crier from reporting if there are prowjobs exist for the same PR that are from a later patchset