Needs rebase: Scan for PRs every six hour instead of 24h
This changes the needs-rebase plugin to scan for open PRs every six
hours instead of the default 24h. This shouldn't be an issue tokenwise,
because the main token usage is for the v4/graphql api plus up to two
tokens per PR that needs rebasing. The number of PRs where we need to
add/remove the label for that will be reduced by running the scan more
often, though.