tests: Skips device plugin for Windows
The test [sig-windows] Device Plugin should be able to create a functioning device plugin for Windows
is supposed to run on nodes with GPUs, which is not a common case, thus, it should
be excluded from most test runs. The test pulls a 10+ GB container image, running
for 100+ minutes, which can lead to test job timeouts.

This has been addressed on master, the test has been tagged with [Feature:GPUDevicePlugin],
and most test jobs are already skipping Feature tests, but that's not the case for
1.21.