kubernetes-csi: fix csi-sanity config in distributed deployment jobs
Overriding the defaults was a mistake, the defaults (pod
csi-hostpathplugin-0, container hostpath) work fine also for
distributed deployment.

The result of this misconfiguration was that paths were created in
/tmp of the socat pod. This accidentally worked in almost all tests
because the driver was too permissive (see
https://github.com/kubernetes-csi/csi-driver-host-path/pull/307) but
failed in the "NodeUnpublishVolume should remove target path" test
because it didn't find the target path in the socat pod.