Branchprotector: Allow to run for a subset of orgs/repos only
Branchprotector consumes a large amount of GitHub api quota, which means
that bigger Prow setups are forced to throttle it which means that it
runs very slowly.

This change allows to run multiple branchprotector instances that each
only manage a subset of org/repos and that each can use different GitHub
users.