kubernetes-csi: more fixes for distributed provisioning
The default csi-hostpathplugin-0 does *not* work for the DaemonSet
that is used for distributed provisioning. Instead, the node and thus
the driver pod are selected through co-scheduling of the socat port
proxy pod. That pod has a predictable name. We can use that to find
the right driver pod.

The necessary support for storing a shell expression in
CSI_PROW_SANITY_POD is only going to be available in
csi-driver-host-path master. For pull requests into
external-provisioner we don't need sanity testing and thus
can avoid the problem.

Some other changes:
- setting CSI_PROW_DRIVER_VERSION is unnecessary when testing
  a PR for csi-driver-host-path
- the Kubernetes version for the periodic canary job gets bumped
  to 1.21 (now supported by KinD 0.11.0) and the job name in
  the dashboard gets fixed (wasn't really "master")