Needs-Rebase: Time out takeAction after five minutes
Needs-Rebase often gets throttled very heavily for hours and more. This
results in takeAction() adding/removing labels and comments based on
outdated data and often multiple times for the same PR.

This change makes takeAction time out after five minutes and goes by the
assumption that the info it's call was based on is likely to be
outdated. We have a periodic rescan as a backstop and assume it will fix
things up if need be.