Branchprotector: Update status contexts if there is at least one optional job
Currently, the branchprotector will ignore repos that do not have
mandatory jobs, unless the repo has an explicit config.
This leads to orphaned required status contexts on GitHub if all jobs on
a repo are made optional.

This change wil make the branchprotector always update the required
status contexts if there is at least one job, optional or not.