config/tests: jobs should not use gs://kubernetes-release-pull
Log policy violations rather than fail the test. Once migration away
from this bucket is complete, the test can be switched to fail instead
of log.

Ideally though, this test would then be switched to a test that fails
for _all_ job types if gs://kubernetes-release-pull shows up as a
substring in _any_ argument.

We could also remove it (possibly N months down the line).