config: add wg-k8s-infra-gcb dashboard to all image-pushing jobs
This involves three sets of changes:
- adding the wg-k8s-infra-gcb dashboard to config/testgrids
- using edit-prowjobs.py to mass-edit all but one image-pushing job
- manually edit post-security-profiles-operator-push-image in
  k8s-staging-seccomp-operator.yaml to add a testgrid-dashboards
  annotation

The last change exposes the fact that there is no one-true-yaml that
defines both prowjobs and all of their testgrid dashboards. The job
has another testgrid dashboard tab defined in a testgrid config. It
can't be moved to the prowjob config as-is, because it has a custom
base_options setting for width, and acontext-free name that wouldn't
make sense on the wg-k8s-infra-gcb dashboard. Neither base_options
nor different tab names are currently supported by the
prowjob-annotation-based testgrid configurator.