blunderbuss: support reviewerCount==0
When the requested number of reviewers is explicitly 0, return no
reviewers.

This allows the plugin to effectively be disabled in one repository
when it is enabled for an entire organization.

Add unit test cases for zero reviewerCount and update existing unit
tests that relied in implicit behavior to set the reviewerCount
explicitly to a non-zero value.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>