Tide status: Handle ReviewApprovedRequired requirement
The status controller didn't have any handling for the
ReviewApprovedRequired requirement and it was only handled in the sync
controller. Until #23296 this happened to work, as the status controller
would only set a PR to success once the sync controller allowed that.
This also meant that PRs that only lacked the review had a "Not
mergeable" status, without any indication what the issue is.