subscriber: use LRU cache
We add a basic LRU (Least Recently Used) cache to try to avoid expensive
lookups (config.GetPresubmits and config.GetPostsubmits). An LRU cache
is basically a smart map that has a set size, and evicts key/value pairs
based on when they were last accessed (the less a key/value pair is
used, the more likely that it will get evicted).

The size of the cache is exposed as a flag in the "sub" command as
"-cache-size <INT>", which by default is set to 1000. Refer to the
cache.go file for an in-depth discussion about the caching
implementation. We try to be as robust as possible in the following
ways:

- we use an off-the shelf cache library (it uses the empty interface{}
  but it's a fair price to pay for robustness)
- if the cache cannot be initialized, we continue without one
- if a cache hit results in a bad type, we throw it away and heal the
  cache entry