releng(kpromo): Drop `--use-service-account` flag
In a previous commit (from PR 23493), the `--use-service-account`
flag was added for `kpromo` jobs, but it might not actually be
necessary (as we're already running in containers with access to
service accounts).

Let's validate this before touching the legacy code paths in CIP
(which handle auth via service account) or building/promoting a
new image containing `gcloud` SDK binaries.

Signed-off-by: Stephen Augustus <foo@auggie.dev>