Move prow image push job to run on tainted high cpu nodes
This job is currently running on default nodes with 8 CPUs in total, I have inspected a running job and found out that it uses 8 CPUs at peak, which might be bounded by the VM. This also affects other jobs running on the same node, including prow components. Make it running on tainted node should help.

I have already created the nodepool with 16 CPUs each and tainted with corresponding key:val