ghproxy: refactor requestCoalescer tests
We tweak the tests to use some more locking mechanisms in the
fakeRequestExecutor. This gives us fine-grained control of deterministic
behaviors without resorting to time.Sleep() for long periods of time. We
no longer have nested goroutines, which makes it easier to reason about
the nature of concurrent threads in this area.

We also make some small changes to the underlying requestCoalescer code,
namely:

1. use an "int" instead of a "bool" for firstRequest.subscribers
2. avoid spurious "defer" keyword in closing the request body in
   requestCoalescer's  RoundTrip()
3. do not intersperse coalescer.Unlock() and firstReq.L.Lock() when
   modifying firstRequest.subscribers.

We also add a new TestStress() which runs the old tests TestRoundTrip and
TestCacheModeHeader 100 times each, all in a few seconds. There is also
a new TestRoundTripSynthetic which adds another concurrency test against
the requestCoalescer.