simplify GetPresubmits and GetPostsubmits type signatures
Previously, if we wanted to get cached versions of presubmits, we had to
pass in not only the cache but also (1) the Config and (2) a
git.ClientFactory. We now include these fields inside the ProwYamlCache
struct directly, so that when we instantiate the cache, we provide
references to these things. Then when it comes time to use the cache, we
just need to look into these fields of the cache.

This has the added benefit of allowing users to know exactly when they
are using the cache vs when they are not, because the type signatures
don't line up with the cached vs non-cached versions of GetPresubmits()
and GetPostsubmits(). Because of this, there is no need for
GetPresubmitsMaybeCached() and GetPostsubmitsMaybeCached(), so we drop
these wrapper functions.