TestGetProwYAMLCachedAndDefaulted: add Env field
The Presets field of ProwYAML is __not__ deep-copied by the
mergePreset() function. Hence this test will only pass if the ProwYAML
itself has been deep-copied already (otherwise, the defaulting logic
will overwrite the cached value's Presets' Env field when the call to
DefaultAndValidateProwYAML() is made, resulting in a write operation
against the cached entry in ProwYAMLCache).

This test only passes because we now use a DeepCopy() call in our
toplevel prowYAMLCache.GetPresubmits() and
prowYAMLCache.GetPostsubmits() functions to the outer ProwYAML object.