sub: do not make ProwYAMLCache a requirement
Most of the time, sub will be called without a Git client because
inrepoconfig will not be used. In these cases, the ProwYAMLCache will
also not be used. So, only exit with a fatal error if (1) the user
wanted to use inrepoconfig (by supplying a GitHub TokenPath), and (2)
the cache initialization failed.