images/bootstrap: allow disabling mtu workaround via env var
This was annoying to roll out to all jobs.  Gate it behind an env var
and default it to true.  We can set the env var to false in the same
preset that sets DOCKER_IN_DOCKER_ENABLED true to roll this out to all
jobs at once without waiting for a bootstrap->bump->kubekins->bump
deployment cycle