config/prow: disable mtu workaround via preset
BOOTSTRAP_MTU_WORKAROUND gates iptables mangling that worked around a
Google Cloud networking issue related to MTU differences in docker and
host networks.  It was necessary for any job that used docker-in-docker.

The underlying issue has since been resolved, so we'd like to disable
the workaround, but leave ourselves the option to re-enable it if the
networking issue reappears.

The workaround itself is in images/bootstrap/runner.sh, which takes
three PRs to change for jobs that use the kubekins image. We are instead
opting to disable the workaround via a config change which takes one PR
to take effect for kubekins-based jobs.

Since any job that needed the workaround was already using the
docker-in-docker present, set the BOOTSTRAP_MTU_WORKAROUND value in the
same preset.

Assuming we see no issues we can default the env var to false in runner.sh
and drop this from the preset, leaving us the option to re-enable via
the preset should the networking issue reappear.