e2e: disable auto-updates on all Fedora CoreOS jobs
This tweaks nodes configurations in order to disable auto-updates
on all jobs running on Fedora CoreOS (FCOS).
An auto-update agent (Zincati) is active by default on all FCOS nodes,
but we don't actually want those machines to perform updates in the
middle of a CI run.
This avoids any possible race between the test-runner and the
machine rebooting itself and switching to a different OS version
during a test run.

Signed-off-by: Luca BRUNO <luca.bruno@coreos.com>