triage: add max_failure_text_length flag
Analyzing a recent failure_data.json has revealed that some failure text
is in the range of 1-30M characters. It's getting truncated down to 10K
for purposes of cluster, but it's still showing up in output as an
example of the failure text being clustered.

This seems excessive, and is probably why failure_data.json is hovering
around 1GB of data uncompressed right now.  I have a strong suspicion
these are all failures from kubernetes/kubernetes integration tests, which
appear to be logging excessively.

Let's cap it to 100K for output, presuming that a human might require more
than 10K of failure text when using triage's UI to see what's going on.