config: export `FinalizeJobConfig` method
Enable experiments with dynamically generating Prowjobs and submitting them to the cluster. It is not hard to generate Prowjob configuration on the fly, but without exporting the finalization method (which is only called on `Load()` otherwise), they do not contain the defaults necessary to be successfuly submitted, such as default decoration configs.