containerd-node-features: skip dynamickubeletconfig tests
DynamicKubeletConfig is not enabled by default (deprecated) and soon to be
removed. Its tests are also incredibly flaky.

As the various changes to the test suite and runner happened, it seems that
containerd's presubmits were missed, Rather than enabling the feature gate
here, lets skip the tests as they're about to be removed and are low signal/high
noise in presubmits.