Fix single override creates many ProwJobs in the system
This commit addresses the problem described in #22690. Override once issued,
can spawn many the almost same prowjobs. The root cause of the problem takes
the origin in the statuses returned by GitHub API. There are multiple statuses
with the same context which reflects the history of the job. One repetition of
context handling for override is enough, rest of the statuses can be ignored.
This commit also adjusts unit tests to the real situation, removing maps as
overlapping statuses are possible.

Signed-off-by: Jakub Guzik <jguzik@redhat.com>