Optimize APIv4 rate limit usage in blunderbuss plugin
This commit fixes random API v4 rate-limit per minute violation when
config option useStatusAvailability is set. In such cases, when a large
set of reviewers is found, there are situations that the client is sending
multiple queries in several seconds, blocking the possibility to query
the data and returning errors for all the left users (code 403). The solution
is to compose a dynamic graphql query that will ask for all the possible reviewers
in two calls (or more when change is huge), thus not firing one call per user
as it is currently implemented.

Signed-off-by: Jakub Guzik <jguzik@redhat.com>