Adds --prepull-images=true to Windows jobs
A new option for E2E tests was introduced in Kubernetes v1.23 [1],
which allows us to prepull some of the most common test images on the
test nodes, like the busybox and agnhost images. Doing this will allow us
to reduce the number of flakes we've been seeing that occur due to tests
timing out (1 minute) while waiting for the images to be pulled.