Crier gerrit reporter: report without URL
Gerrit adaptor depends on comments posted by crier gerrit reporter to figure out which prow jobs to be triggered with /retest. #24770 truncates gerrit reports when there are too many prow jobs, resulted in not all prow jobs are reported in the comments. So when a user comments /retest prow only triggers the jobs that were included in the report, instead of all prow jobs that failed.