kubernetes-sigs/nfd: verify job for old release branches
Create a separate verify job for older release branches that are based
on Golang v1.16 or earlier. These branches started to fail (gofmt) after
the verify job was bumped to use Golang v1.17. This patch is just in
case we submit/cherry-pick some changes to the old release branches - to
avoid using /override prow commands.