sidecar/censoring: allow opting into fast failure
If a user has set up secret censoring on output and knows that it should
succeed in most cases, it may be useful for them to stop treating the
censoring process as best-effort. In such a case, failures during
censoring would cause the whole `sidecar` process to crash. This would
lead to terrible UX for the user of an affected job, but we can do no
better, as even uploading the job logs may be unsafe in such a
situation.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>