This commit adds a new job `ci-kubernetes-node-kubelet-lock-contention` to the test suite.
This test is created due to the need of `--lock-file` and
`--exit-on-lock-contention` be moved as flags and into the Kubelet
configuration file rather than be dropped.

Adds a new tab in TestGrid named `kubelet-gce-e2e-lock-contention`
running tests focused on `NodeSpecialFeature:LockContention`.

Corresponding e2e test is at https://github.com/kubernetes/kubernetes/pull/104334

Signed-off-by: Imran Pochi <imranpochi@microsoft.com>