clean up infrastructure for beta API e2e tests
Instead of using a testing tag [Beta] which would require the test name be
updated during a transition to GA and doesn't directly match the state
of featuregate itself or handle beta fields gracefully, we can instead
use the discovery API in the e2e test itself to decide whether or not to
skip a test.  The same way that we check inside of the
kube-controller-manager.

Looking at the results of the new jobs in prow, I've realized that to
start, we only need to have a single job that looks like the default job
except that the kube-apiserver enables all beta APIs.