Prow image build presubmit job runs on high cpu node
Noticed that the image build job takes ~16 minutes normally, could be up to 25 minutes. The CPU profile from GKE reveals that the job uses all the CPU from a 8 cpu node. Updating the job so that it only runs on a nodepool with highcpu label.

/hold
the nodepool doesn't exist yet