Prow jobs respect default entrypoints
Supercedes https://github.com/kubernetes/test-infra/pull/25383, where in #25383 the logic is in the pod where tests run, the problem there is that imagePullSecret defined on job pod is not accessible, which limits it's use case. The other problem is that the containers inside a test pod are not aware of the image they are from, so it's a little bit weird to let them know about it.

This PR moves the logic to plank, which feels more natural

-----------------------------
This is an unfortunate fact of prow, that user need to explictly set the entrypoint.

The migration of prow images from being built with bazel to ko introduced a side effect of all prow jobs that use gcr.io/k8s-prow images, such as robots/comment, robots/pr-creator etc. would fail due to the location of default entrypoint change. It would be trivial amount of work to update the binary location in prow jobs definition, but would like to use this opportunity to try to get this fixed.