jira, fakejira: add new functions to the jira client
This PR adds a lot of new functionality to the JIRA client that will be
required for the changes being made to the JIRA prow plugin. The new
functions have been tested on a live JIRA instance to verify that they
properly work.

This PR also contains an updated fake client to match the new Client
interface definition. The fake client is untested at the moment.  Unit
tests to ensure functionality will come in a later PR.