Introduce max delay time to throttling algorithm
This commit introduces max delay time to the additional throttling algorithm.
When there is an issue with gh api access, an algorithm that was supposed to work
locally (in timespan of around 10-60 seconds per queue) is co-responsible for forming
a massive queue of messages, scheduled to be handled in the future. This is unwanted
behavior. When there is an external problem and the bigger queue forms, the algorithm will
use max delay time provided by the user to avoid forming massive queue of messages.

Signed-off-by: Jakub Guzik <jguzik@redhat.com>