Tide: Consider PRs that are being serially retested for batches
Currently, Tide requires a PR to have passing required contexts to
include it in a batch. This changes that to also allow pending context
if there is a prowjob for the pending context that was created by Tide.

We use a label to detect if Tide created a given ProwJob.