crier: gerrit: actually respect commentSizeLimit
Previously, we could generate a comment slightly over the
`commentSizeLimit` if we ran into errors because we appended an error
line at the end. This is no longer the case as we take into account the
size of this possible error line as well during GenerateReport().

As a result, the commentSizeLimit field of the unit tests have to take
into account the length of the error messages as well (this is why these
fields have higher values now).