releng: Remove configs for k8s-beta jobs / add k8s-stable4 variants
For years, we've had issues with the `k8s-beta` generic version
marker.

There are a few scenarios...

During the dev cycle:
- beta should not exist, but actually points to the previous release

After a beta release:
- beta should exist, but:
  1. no one really updated this manually
  2. the branch cuts and job creation happens at the RC stage

After an RC release (branch cut):
- beta should have already existed
- beta represents the RC version after branch job creation

After an official release:
- beta should not exist, but actually points to the previous release

None of these scenarios are acceptable, so it's time to remove this
marker. Generated jobs that reference `beta` in their name have
already been using version-specific markers for multiple releases, so
this has minimal impact.

Here we instead add a `k8s-stable4` marker, to provide a version
marker for a release branch that is in maintenance mode/soon-to-be
EOL-ed.

Adding a marker at the end of the set means we no longer have to
consider moving the `k8s-beta` marker during a release cycle.

Signed-off-by: Stephen Augustus <foo@auggie.dev>