GitHub: Fix apps auth to work with base urls that contain a path
Currently, apps auth does not work with GitHub base urls that include a
path, which is used e.G. in GitHub Enterprise setups.

This is because the apps auth roundtripper decides how to authenticate a
given request based on the path. In order for that to work correctly,
the parts of the path that come from the base url need to be stripped
first.

This in turn requires to parse all base urls in order to safely extract
the path, if any. That results in the GitHub client creation now
potentially erroring, which is why this change got bigger. Changing this
is a good thing though, as it moves a fatal error condition from runtime
to client creation time (doing requests against an invalid base URL will
never work).