Job history page: use a gray color for aborted jobs
The job history page uses a yellow color for "ABORTED" jobs, this is the
same color as the one used for pending jobs. This similarity is
confusing as it makes aborted jobs look as if they're still running.

This commit makes it so that aborted jobs will get their own color,
gray, to indicate that they have been aborted. This color was chosen
because its neutral - aborted jobs are not very interesting, they don't
indicate a failure or a success, they're usually a result of a PR being
forced-pushed to or a PR being closed while the job was running.

Signed-off-by: Omer Tuchfeld <omer@tuchfeld.dev>