node-swap-fedora: Fix ignition
node-swap-fedora and node-swap-fedora-serial have been failing for >20
days as crio did not become ready before the test timed out and failed.

This was because the crio-install service was being interrupted by the
dbus-tools installation restart.

Here we move the crio-swap jobs inline with other crio jobs that install
dbus and install-in-place rather than doing a reboot.

If for some reason these jobs are special and need the reboot path,
we'll be able to revert to the existing method and use systemd dependencies
to wait for the dbus restart instead.