Adds private registry preset for Windows GCE jobs
The test [sig-node] Container Runtime blackbox test when running a container with a new image should be able to pull from private registry with secret [NodeConformance]
is testing that images from private registries can be pulled using a
docker config file containing the right credentials.

The preset preset-windows-private-registry-cred contains the necessary
secrets and env variables that allows the test to pass on Windows, providing
it the necessary credentials to pull from a registry that contains a Windows
manifest list containing an image for Windows 20H2.