windows: Adds -prepull-images=true for a few Windows jobs
Windows images tend to be larger than the Linux ones. Some tests expect
the pod to become active in 1 minute, which might not be enough to pull
the image and start the container, resulting in flaky tests.

Most of the Windows test runs uses this flag to prepull some of the commonly
used test images, like busybox and agnhost.