Enable auto fast-forward job in nomock mode
We have proven that the job works as expected in mock mode (without
pushing to upstream) during the past release cycle.

This means we now can promote the jobs to `sig-release-releng-blocking`
as well as enable it finally.

After that we will update the testfreeze plugin to mention the new
process, adapt docs as well as sending a not to the k-dev mailing list.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>