cherrypicker better support for cherry-pick within a cherry-pick PR
# Change tl;dr

Avoid creating very long titles when doing a cherry-pick of a
cherry-pick

# Motivation

In OpenShift repositories it's easier to cherry-pick to multiple
versions "serially" rather than all from the main PR. What I mean by
that is that we first `/cherry-pick X` in the master branch PR, then
once a new cherry pick PR is created for the X branch, we submit another
`/cherry-pick X-1` comment in the X PR, rather than in the master PR. We
do that because we have another bot that handles a lot of boring
BugZilla work automatically for us, and it's crucial for that bot that
those cherry-picks happen serially.

This almost works OK, except for a small issue with the PR titles that
this commit aims to fix.

# The issue

When asking the bot to do a cherry-pick when already in a branch that
was created (and more importantly, titled) by the bot, the new
cherry-pick PR will have a title with a target branch indicator
containing both the indicator of the original cherry-pick, and of
the target branch of the new cherry-pick. This is undesired as
a sequence of cherry-pick would create ever-growing titles unless
the user manually trims the titles to remove older indicators.

Here is an example demonstrating the issue:
```
Original PR title: "Hello world"
Backport to release-9.9 title: "[release-9.9] Hello world"
Backport to release-9.8 title: "[release-9.8] [release-9.9] Hello world"
```

And here is how it would look like after the fix:
```
Original PR title: "Hello world"
Backport to release-9.9 title: "[release-9.9] Hello world"
Backport to release-9.8 title: "[release-9.8] Hello world"
```

# The fix

See the comment above `omitBaseBranchFromTitle` for details, but in
short, when creating the title for a cherry-pick, we should also
consider the target branch of the PR the cherry-pick comment originated
from (henceforth "Base branch").

We then make sure, when creating the title for the new cherry-pick PR,
to remove the existing `[base branch] ` indicator, if one exists, so
that when we add `[cherry-pick branch] ` it would be the only indicator.