prow/plugin: Extends label plugin to use Unique Prefixes
This commit takes care of label prefixes that can only
exist once per issue. This is done by using a list of
unique prefixes to create a map that maps these prefixes
to existing labels that match the prefix. When a new
label is needed to be added that matches a unique
prefix, the pre-existing labels in this map that match
the same prefix are removed and then the new label is
added.

Signed-off-by: Raghav Roy <raghavroy145@gmail.com>