Fix not welcoming members when alwaysPost=true
Welcome plugin configuration of ``always_post: true`` is not always
respected, because when user is trusted we're short-circuiting the
function and not posting the message.

This change keeps the short-circuit branch so that we won't need to
render the message template if not needed, but only doing so if not
configured to always post the message.

Tests were changed to use modified trigger plugin configuration of
``only_org_members``, to make actual difference in the way we perceive
trusted entities and nontrusted ones.