integration tests: sub: add test case for cloning from inrepoconfig
The fake (nonsense) cookie file we mount is only there to trigger the
code branch that detects whether to create a git client factory for
GitHub or for Gerrit. The --cookie-file flag makes sub's ProwYAML cache
initialization logic assume that all inrepoconfigs are cloned from
Gerrit.

For our purposes the contents of the cookie file
doesn't matter because we don't even hit the fakegerritserver URL for
cloning the repo because we use a direct CloneURI field.

Note that this tightly couples this deployment of sub to Gerrit (see the
TODO comment added in sub.yaml).