jira: general client improvements
This PR fixes the fakejira `CloneIssue` and `CreateIssue` functions
along with adding unit tests for them, adds more extensive issue
function errors for the fakeclient, and modifies the
`DeleteRemoteLinkViaURL` function to return a `bool` indicating whether
a link with the provided URL was found and removed (no remote link with
the provided URL is not an error).

These improvements will be used in a new openshift external prow plugin
called the `jira-lifecycle` plugin, which will replicate the `bugzilla`
plugins functionality, but with the Jira service instead of bugzilla.