label_sync: Add new global label - area/community-meeting
Adding this label is a first step towards automating the
process of curating content for the K8s community meeting.
This label can be added by anyone and before a community
meeting, the host/set of people responsible for the meeting
filter through these issues and/or PRs and curate the content.

Signed-off-by: Madhav Jivrajani <madhav.jiv@gmail.com>