kubernetes-csi: more realistic pre-merge testing of csi-release-tools
In https://github.com/kubernetes-csi/csi-release-tools/pull/195, the
csi-driver-host-path version was bumped while keeping the default Kubernetes
version at the (ancient!) v1.17.0. When updating csi-release-tools later in
csi-test, pull-kubernetes-csi-csi-test failed because the new driver is not
supporting Kubernetes 1.17 anymore (no v1 CSIDriver API).

This wasn't detected by the csi-release-tools pre-submmit because it overwrote
the Kubernetes version. That's okay for other repos (they do the same in their
jobs) but not for csi-test.