Gerrit client accepts org/repo from different sources
This was overlooked in previous PR, that the Gerrit org/repo configuration is actually defined at different places in prow configuration. Make it flexible so that it'll work with tide in the future.

Also fixing a bug that when LastSyncTracker is supplied as nil, it won't panic(Crier would panic without this fix)