releng: consolidate alphafeatures settings
KUBE_FEATURE_GATES was changed when v1.22 was stable1, but it was never
propagated back. The stable3 and stable2 jobs are breaking now.

The history behind --test_args for alphafeatures jobs is murkier since
it predates the move to releng. Suffice to say it's been tribally edited
and manually copied around since before there was a --test_args flag in
2017. It should have made it to stable3 sometime in 2018.

At first I copied these settings back from stable1 to stable2 to
stable3, but if they're all the same now, we can consolidate to avoid
the copy paste