sig node: add presubmit test for Dynamic Resource Allocation
No code has been merged yet, but
https://github.com/kubernetes/kubernetes/pull/111023 will add it and contains
E2E tests that need to pass before merging.

The job is configured so that it only runs if relevant code is getting modified
by a PR. The tests are new and therefore cannot be merge-blocking because that
might prevent getting relevant changes merged.