Crier reports started.json with commit hash
The commit hash is the final SHA resulted from cloning, it's not available until clonerefs is finished.  was uploaded by Crier when prowjob CR is just created, which is even ahead of job pod is scheduled, so there is no chance clone is done. However, when prowjob is completed the final SHA could be derived from clone record that is uploaded to GCS