Deck: Fix colors in job history view
When the uploading of finished.json was moved to crier, it also ended up
inserting a lowercased result. This breaks the coloring in the template,
as that expects it to be uppercased.

As there is data following both patterns out there already, fix this in
Deck by uppercasing the result field before passing it onto the
template.