config: Auto close issues as Not Planned
This change makes use of the `note_planned` state of the GH API
to auto close issues. Please note that this state is only for issues
and not PRs.

Due to this, the current triage bot is split into 2: one for issues
(which now comments `/close not-planned`) and one for PRs (which will
still comment `/close`).

Signed-off-by: Madhav Jivrajani <madhav.jiv@gmail.com>