plugins/lifecycle: Better handle close regexes
This change better handles close regexes by checking and
ensuring that PRs cannot be closed as Not Planned. The
reason for this is that the "not_planned" state does not
exist for PRs, but only exists for issues which is why
allowing a PR to be closed as by commenting `/close not-planned`
feels awkward.

This change also adds the requisite tests for the same.

Signed-off-by: Madhav Jivrajani <madhav.jiv@gmail.com>