NFD: Restructure jobs based on the branch version
This patch restructures jobs per branch so that we can control
parameters differently depending on the branch name. Starting from
release-0.10 we will have a separate file (branch name as prefix)
to keep branch specific parameters. For those jobs that are common
between multiple branches, a generic file is created.

Signed-off-by: Feruzjon Muyassarov <feruzjon.muyassarov@intel.com>